moved saveCollectionInBackground function to AnkiActivity Because it was not proper to have a function inside UIUtils that schedules some backgroud operations associated with collection
