+++ AnkiDroid/src/main/java/com/ichi2/anki/NoteEditor.kt
@@ -459,7 +459,7 @@ class NoteEditor : AnkiActivity(), DeckSelectionListener, SubtitleListener, Tags
         super.onStop()
         if (!isFinishing) {
             WidgetStatus.update(this)
-            UIUtils.saveCollectionInBackground()
+            saveCollectionInBackground()
         }
     }
